Restore uc_flags to unsigned long in ucontext
struct ucontext in <sys/ucontext.h> is used for user-level context
and is kernel indenpendent. It is separate from struct ucontext
defined in kernel header files.

+++ sysdeps/unix/sysv/linux/x86_64/sys/ucontext.h
@@ -132,7 +132,7 @@ typedef struct
 /* Userlevel context.  */
 typedef struct ucontext
   {
-    unsigned long long int uc_flags;
+    unsigned long int uc_flags;
     struct ucontext *uc_link;
     stack_t uc_stack;
     mcontext_t uc_mcontext;
